[7418] Create commonDescriptors to reuse
