+++ ui/app/src/components/ContentTypeManagement/descriptors/controls/dateTime.ts
@@ -17,6 +17,7 @@
 import { createVirtualSection, DescriptorContentType } from '../../utils';
 import { immutableEmptyObject } from '../../../../utils/object';
 import { defineMessage } from 'react-intl';
+import { commonFieldPropertiesDescriptors } from './commonDescriptors';
 
 export const dateTimeDescriptor: DescriptorContentType = {
 	id: 'date-time',
@@ -102,13 +103,7 @@ export const dateTimeDescriptor: DescriptorContentType = {
 			defaultValue: undefined,
 			validations: immutableEmptyObject
 		},
-		readonly: {
-			id: 'readonly',
-			type: 'boolean',
-			name: defineMessage({ defaultMessage: 'Read Only' }),
-			defaultValue: undefined,
-			validations: immutableEmptyObject
-		},
+		readonly: commonFieldPropertiesDescriptors['readonly'],
 		readonlyEdit: {
 			id: 'readonlyEdit',
 			type: 'boolean',
@@ -116,13 +111,7 @@ export const dateTimeDescriptor: DescriptorContentType = {
 			defaultValue: undefined,
 			validations: immutableEmptyObject
 		},
-		required: {
-			id: 'required',
-			type: 'boolean',
-			name: defineMessage({ defaultMessage: 'Required' }),
-			defaultValue: undefined,
-			validations: immutableEmptyObject
-		}
+		required: commonFieldPropertiesDescriptors['required']
 	},
 	supportedPostFixes: ['_dt']
 };
