+++ ui/app/src/components/ContentTypeManagement/descriptors/controls/rte.ts
@@ -17,6 +17,7 @@
 import { createValidation, createVirtualSection, DescriptorContentType } from '../../utils';
 import { immutableEmptyObject } from '../../../../utils/object';
 import { defineMessage } from 'react-intl';
+import { commonFieldPropertiesDescriptors } from './commonDescriptors';
 
 export const rteDescriptor: DescriptorContentType = {
 	id: 'rte',
@@ -108,13 +109,7 @@ export const rteDescriptor: DescriptorContentType = {
 				type: createValidation('type', 'item')
 			}
 		},
-		required: {
-			id: 'required',
-			type: 'boolean',
-			name: defineMessage({ defaultMessage: 'Required' }),
-			defaultValue: undefined,
-			validations: immutableEmptyObject
-		}
+		required: commonFieldPropertiesDescriptors['required']
 	},
 	supportedPostFixes: ['_html']
 };
