+++ ui/app/src/components/ContentTypeManagement/descriptors/controls/transcodedVideoPicker.ts
@@ -15,8 +15,8 @@
  */
 
 import { createValidation, createVirtualSection, DescriptorContentType } from '../../utils';
-import { immutableEmptyObject } from '../../../../utils/object';
 import { defineMessage } from 'react-intl';
+import { commonFieldPropertiesDescriptors } from './commonDescriptors';
 
 export const transcodedVideoPickerDescriptor: DescriptorContentType = {
 	id: 'transcoded-video-picker',
@@ -44,20 +44,8 @@ export const transcodedVideoPickerDescriptor: DescriptorContentType = {
 				type: createValidation('type', 'transcoded-video')
 			}
 		},
-		readonly: {
-			id: 'readonly',
-			type: 'boolean',
-			name: defineMessage({ defaultMessage: 'Read Only' }),
-			defaultValue: undefined,
-			validations: immutableEmptyObject
-		},
-		required: {
-			id: 'required',
-			type: 'boolean',
-			name: defineMessage({ defaultMessage: 'Required' }),
-			defaultValue: undefined,
-			validations: immutableEmptyObject
-		}
+		readonly: commonFieldPropertiesDescriptors['readonly'],
+		required: commonFieldPropertiesDescriptors['required']
 	},
 	supportedPostFixes: ['_o']
 };
